requirements.txt
pyaudio
websockets
python-dotenv
fish-audio-sdk
